// API/models/diet.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Appei
 * API of Appetee Project
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface Diet
 */
export interface Diet {
    /**
     * 
     * @type {number}
     * @memberof Diet
     */
    id?: any;
    /**
     * 
     * @type {string}
     * @memberof Diet
     */
    nom: any;
}