// API/models/diet-user.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Appei
 * API of Appetee Project
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface DietUser
 */
export interface DietUser {
    /**
     * 
     * @type {boolean}
     * @memberof DietUser
     */
    allergy: any;
    /**
     * 
     * @type {number}
     * @memberof DietUser
     */
     id_diet?: any;
}