// API/models/ingredients.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Appei
 * API of Appetee Project
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface Ingredients
 */
export interface Ingredients {
    /**
     * 
     * @type {number}
     * @memberof Ingredients
     */
    id?: any;
    /**
     * 
     * @type {string}
     * @memberof Ingredients
     */
    nom: any;
    /**
     * 
     * @type {string}
     * @memberof Ingredients
     */
    unit?: any;
    /**
     * 
     * @type {string}
     * @memberof Ingredients
     */
     ingredient_type?: any;
}