// API/models/user.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Appei
 * API of Appetee Project
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface User
 */
export interface User {
    /**
     * 
     * @type {number}
     * @memberof User
     */
    id?: any;
    /**
     * 
     * @type {string}
     * @memberof User
     */
    surname: any;
    /**
     * 
     * @type {string}
     * @memberof User
     */
    email?: any;
    /**
     * 
     * @type {string}
     * @memberof User
     */
     password_user?: any;
    /**
     * 
     * @type {boolean}
     * @memberof User
     */
     is_child?: any;
    /**
     * 
     * @type {string}
     * @memberof User
     */
     image_user_url?: any;
    /**
     * 
     * @type {Array&lt;IngredientUser&gt;}
     * @memberof User
     */
     disliked_ingredient?: any;
    /**
     * 
     * @type {Array&lt;DietUser&gt;}
     * @memberof User
     */
    diets?: any;
    /**
     * 
     * @type {Array&lt;User&gt;}
     * @memberof User
     */
     child_user?: any;
}